+++ spec/rubocop/cop/style/optional_boolean_parameter_spec.rb
@@ -42,9 +42,16 @@ RSpec.describe RuboCop::Cop::Style::OptionalBooleanParameter do
     RUBY
   end
 
-  it 'does not register an offense when defining method with optonal non-boolean arg' do
+  it 'does not register an offense when defining method with optional non-boolean arg' do
     expect_no_offenses(<<~RUBY)
       def some_method(bar = 'foo')
+      end
+    RUBY
+  end
+
+  it 'does not register an offense when defining respond_to_missing? method with boolean arg' do
+    expect_no_offenses(<<~RUBY)
+      def respond_to_missing?(arg, bar = false)
       end
     RUBY
   end